Chore: Rename discover vue file to Discover.vue
